4565: Started work on upgrading redux-toolkit
