+++ assets/admin/components/activation-code/activation-code-create.jsx
@@ -1,7 +1,7 @@
 import { React, useState, useEffect } from "react";
 import { useTranslation } from "react-i18next";
 import { useNavigate } from "react-router-dom";
-import { usePostV2UserActivationCodesMutation } from "../../redux/api/api.generated.ts";
+import { usePostV2UserActivationCodesMutation } from "../../../shared/redux/generated-api.ts";
 import ActivationCodeForm from "./activation-code-form";
 import {
   displaySuccess,
