+++ assets/admin/components/feed-sources/feed-sources-list.jsx
@@ -5,7 +5,7 @@ import {
   useGetV2FeedSourcesQuery,
   useDeleteV2FeedSourcesByIdMutation,
   useGetV2FeedSourcesByIdSlidesQuery,
-} from "../../redux/api/api.generated.ts";
+} from "../../../shared/redux/generated-api.ts";
 import ListContext from "../../context/list-context";
 import ContentBody from "../util/content-body/content-body";
 import List from "../util/list/list";
