+++ assets/admin/components/groups/group-create.jsx
@@ -1,7 +1,7 @@
 import { React, useState, useEffect } from "react";
 import { useTranslation } from "react-i18next";
 import { useNavigate } from "react-router-dom";
-import { usePostV2ScreenGroupsMutation } from "../../redux/api/api.generated.ts";
+import { usePostV2ScreenGroupsMutation } from "../../../shared/redux/generated-api.ts";
 import GroupForm from "./group-form";
 import {
   displaySuccess,
