+++ assets/admin/components/media/media-create.jsx
@@ -1,6 +1,6 @@
 import { React, useState, useEffect } from "react";
 import { useTranslation } from "react-i18next";
-import { usePostMediaCollectionMutation } from "../../redux/api/api.generated.ts";
+import { usePostMediaCollectionMutation } from "../../../shared/redux/generated-api.ts";
 import MediaForm from "./media-form";
 import {
   displayError,
