+++ assets/admin/components/playlist/campaign-form.jsx
@@ -2,7 +2,7 @@ import { React } from "react";
 import { useTranslation } from "react-i18next";
 import PropTypes from "prop-types";
 import idFromUrl from "../util/helpers/id-from-url";
-import { useGetV2CampaignsByIdScreenGroupsQuery } from "../../redux/api/api.generated.ts";
+import { useGetV2CampaignsByIdScreenGroupsQuery } from "../../../shared/redux/generated-api.ts";
 import ContentBody from "../util/content-body/content-body";
 import SelectScreensTable from "../util/multi-and-table/select-screens-table";
 import SelectGroupsTable from "../util/multi-and-table/select-groups-table";
