+++ assets/admin/components/playlist/playlist-campaign-edit.jsx
@@ -3,7 +3,7 @@ import { useParams } from "react-router-dom";
 import PropTypes from "prop-types";
 import PlaylistCampaignManager from "./playlist-campaign-manager";
 import idFromUrl from "../util/helpers/id-from-url";
-import { useGetV2PlaylistsByIdQuery } from "../../redux/api/api.generated.ts";
+import { useGetV2PlaylistsByIdQuery } from "../../../shared/redux/generated-api.ts";
 
 /**
  * The playlist/campaign edit component.
