+++ assets/admin/components/screen/screen-edit.jsx
@@ -2,7 +2,7 @@ import { React, useEffect, useState } from "react";
 import { useParams } from "react-router-dom";
 import idFromUrl from "../util/helpers/id-from-url";
 import ScreenManager from "./screen-manager";
-import { useGetV2ScreensByIdQuery } from "../../redux/api/api.generated.ts";
+import { useGetV2ScreensByIdQuery } from "../../../shared/redux/generated-api.ts";
 
 /**
  * The screen edit component.
