+++ assets/admin/components/slide/content/feed-selector.jsx
@@ -7,7 +7,7 @@ import set from "lodash.set";
 import {
   api,
   useGetV2FeedSourcesQuery,
-} from "../../../redux/api/api.generated.ts";
+} from "../../../../shared/redux/generated-api.ts";
 import MultiSelectComponent from "../../util/forms/multiselect-dropdown/multi-dropdown";
 import idFromUrl from "../../util/helpers/id-from-url";
 import ContentForm from "./content-form";
