+++ assets/admin/components/util/multi-and-table/select-playlists-table.jsx
@@ -6,7 +6,7 @@ import {
   useGetV2PlaylistsQuery,
   useGetV2SlidesByIdPlaylistsQuery,
   useGetV2PlaylistsByIdSlidesQuery,
-} from "../../../redux/api/api.generated.ts";
+} from "../../../../shared/redux/generated-api.ts";
 import PlaylistsDropdown from "../forms/multiselect-dropdown/playlists/playlists-dropdown";
 import { SelectPlaylistColumns } from "../../playlist/playlists-columns";
