+++ assets/admin/index.jsx
@@ -2,7 +2,7 @@ import React from "react";
 import { BrowserRouter } from "react-router-dom";
 import { createRoot } from "react-dom/client";
 import { Provider } from "react-redux";
-import { store } from "./redux/store";
+import { store } from "../shared/redux/store.js";
 import App from "./app";
 
 const container = document.getElementById("root");
