--- assets/admin/redux/README.md
@@ -1,16 +1,0 @@
-# Redux setup
-
-The Redux setup is built with Redux Toolkit and RTK Query for communicating with the api.
-
-The api integration is generated with <https://github.com/rtk-incubator/rtk-query-codegen>.
-
-The folder `src/redux/api` contains the latest generated ApiSlice and a node setup for generating a new ApiSlice:
-
-```shell
-# Override api.json with new OpenAPI specification.
-
-npm install
-node start
-```
-
-The generated api is typescript which is compiled to a javascript file (`api.generated.js`).
