--- assets/admin/redux/api/.gitignore
@@ -1,1 +1,0 @@
-node_modules/
