--- assets/admin/redux/api/package.json
@@ -1,20 +1,0 @@
-{
-  "name": "api-generation",
-  "version": "1.0.0",
-  "description": "Provides project for generating api from OpenAPI spec and converting from typescript to vanilla javascript.",
-  "main": "index.js",
-  "scripts": {
-    "generate": "npx @rtk-incubator/rtk-query-codegen-openapi --hooks api.json > api.generated.ts",
-    "//": "# Therefore we replace into a temporary file which is then renamed to the original file",
-    "replace": "sed 's/import { fetchBaseQuery } from .*;/import extendedBaseQuery from \"..\\/dynamic-base-query\";/g;s/baseQuery: fetchBaseQuery.*,/baseQuery: extendedBaseQuery,\\n  keepUnusedDataFor: 0,/g' api.generated.ts > api.generated.ts.patched && mv api.generated.ts.patched api.generated.ts",
-    "compile": "node_modules/typescript/bin/tsc api.generated.ts",
-    "start": "npm run generate && npm run replace"
-  },
-  "author": "ITKDev",
-  "license": "ISC",
-  "dependencies": {
-    "@reduxjs/toolkit": "^1.6.1",
-    "@rtk-incubator/rtk-query-codegen-openapi": "^0.5.1",
-    "typescript": "^4.4.3"
-  }
-}
