+++ assets/shared/redux/store.js
@@ -1,5 +1,5 @@
 import { configureStore } from "@reduxjs/toolkit";
-import { api } from "./api/api.generated.ts";
+import { api } from "./generated-api.ts";
 
 /* eslint-disable-next-line import/prefer-default-export */
 export const store = configureStore({
