+++ package.json
@@ -9,9 +9,10 @@
   },
   "devDependencies": {
     "@playwright/test": "1.53.2",
-    "@reduxjs/toolkit": "^1.6.1",
+    "@rtk-query/codegen-openapi": "^2.0.0",
     "@vitejs/plugin-react": "^4.6.0",
     "esbuild": "^0.25.8",
+    "esbuild-runner": "^2.2.2",
     "sass": "^1.88.9",
     "typescript": "^4.4.2",
     "vite": "npm:rolldown-vite@^7.0.3",
@@ -25,6 +26,7 @@
     "@fortawesome/react-fontawesome": "^0.2.2",
     "@hello-pangea/dnd": "^16.0.0",
     "@popperjs/core": "^2.11.8",
+    "@reduxjs/toolkit": "^2.8.2",
     "@u-wave/react-vimeo": "^0.9.11",
     "@vitejs/plugin-react-oxc": "^0.3.0",
     "bootstrap": "^5.3.7",
